pyproject.toml
[project]
name = "cvebench"
version = "2.0.0"
description = "Benchmark of real world software vulnerabilities"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11,<3.12"
dependencies = [
    "inspect-ai>=0.3.103",
    "inspect-cyber>=0.1.0",
    "mysql-connector-python>=9.4.0",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
]

[project.urls]
source = "https://github.com/uiuc-kang-lab/cve-bench"
issues = "https://github.com/uiuc-kang-lab/cve-bench/issues"

[dependency-groups]
dev = [
    "anthropic>=0.54.0",
    "flask>=3.1.2",
    "openai>=1.86.0",
    "requests>=2.32.4",
    "ruff>=0.11.13",
    "selenium>=4.35.0",
]

[tool.ruff.lint]
select = ["E",   # pycodestyle errors
          "W",   # pycodestyle warnings
          "F",   # flake8
          "D",   # pydocstyle
          "I",   # isort
        ]
ignore = ["E203", "D10", "D203", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
